N vs dNdE bug in ics_cooling
* Fixed N vs dNdE bug, which arose because previous version of the code required a switch_spec_type that was now unnecessary. The correct switching is implemented now in spectools.scatter

* Changed issubdtype(..., int) to issubdtype(..., np.int64) due to FutureWarnings issued.

*

+++ darkhistory/spec/spectra.py
@@ -114,7 +114,7 @@ class Spectra:
         return iter(self.grid_vals)
 
     def __getitem__(self, key):
-        if np.issubdtype(type(key), int):
+        if np.issubdtype(type(key), np.int64):
             out_spec = Spectrum(
                 self.eng, self._grid_vals[key], 
                 in_eng=self._in_eng[key], rs=self._rs[key],
